Update: test sobre sa-file

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -5,35 +5,67 @@ import { SaveFile } from './save-file.use-case';
 
 
 describe('SaveFileUseCase', () => {
+    // al ser lo unico obligatorio fileContent
+    const customOptions = {
+        fileContent: 'custom content',
+        fileDestination: 'custom-outputs/file-destination',
+        fileName: 'custom-table-name'
+    }
+    // path a evaluar
+    // file Path
+    const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
 
     // borrar la carpeta despues de cada prueba
     afterEach(() => {
+
+        const outputFolderExist = fs.existsSync('outputs');
         // borrar el archivo: clean Up
-        fs.rmSync('outputs', { recursive: true });
+        if ( outputFolderExist ) fs.rmSync('outputs', { recursive: true });
+
+        const customOutputFolderExist = fs.existsSync(customOptions.fileDestination);
+        // borrar el archivo: clean Up
+        if ( customOutputFolderExist ) fs.rmSync(customOptions.fileDestination, { recursive: true });
+        
     });
 
     test('should save file with default values', () => {
         // creamos la instancia de la clase SaveFile()
         const saveFile = new SaveFile();
         // file Path
-        const filePath = 'outputs/table.txt';
+        const customFilePath = 'outputs/table.txt';
         // al ser lo unico obligatorio fileContent
-        const options = {
+        const customOptions = {
             fileContent: 'test content'
         }
 
         // creamos un resultado y mandamos las opciones
-        const result = saveFile.execute(options);
+        const result = saveFile.execute(customOptions);
                 
-        const fileExists = fs.existsSync(filePath);// aqui debemos tener cuidado porque podria dar un falso positivo
-        const fileContent = fs.readFileSync(filePath, { encoding: 'utf-8' });
+        const fileExists = fs.existsSync(customFilePath);// aqui debemos tener cuidado porque podria dar un falso positivo
+        const fileContent = fs.readFileSync(customFilePath, { encoding: 'utf-8' });
 
         
         // expect
         // experamos que sea un true
         expect( result ).toBe( true );
         expect( fileExists ).toBe( true );
-        expect( fileContent ).toBe( options.fileContent );
+        expect( fileContent ).toBe( customOptions.fileContent );
+    });
+
+    test('should save file with custom values', () => {
+        // creamos la instancia de la clase SaveFile()
+        const saveFile = new SaveFile();
+        
+        const result = saveFile.execute( customOptions );
+        const fileExists = fs.existsSync( customFilePath );
+        const fileContent = fs.readFileSync(customFilePath, { encoding: 'utf-8' });
+
+
+        // expects
+        expect( result ).toBe( true );
+        expect( fileExists ).toBe( true );
+        expect( fileContent ).toBe( customOptions.fileContent );
+
     });
 
 });
